Connect with permission

+++ src/main/kotlin/edu/ingsis/snippetmanager/snippet/SnippetService.kt
@@ -1,6 +1,7 @@
 package edu.ingsis.snippetmanager.snippet
 
 import edu.ingsis.snippetmanager.external.asset.AssetApi
+import edu.ingsis.snippetmanager.external.permission.PermissionService
 import edu.ingsis.snippetmanager.external.printscript.PrintScriptApi
 import edu.ingsis.snippetmanager.external.printscript.dto.ValidateDTO
 import edu.ingsis.snippetmanager.snippet.dto.CreateSnippetDto
@@ -9,6 +10,7 @@ import edu.ingsis.snippetmanager.snippet.dto.SnippetDto
 import jakarta.transaction.Transactional
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
+import org.springframework.security.oauth2.jwt.Jwt
 import org.springframework.stereotype.Component
 import org.springframework.web.multipart.MultipartFile
 import org.springframework.web.server.ResponseStatusException
@@ -20,51 +22,60 @@ class SnippetService
         private val repository: SnippetRepository,
         private val printScriptService: PrintScriptApi,
         private val assetService: AssetApi,
+        private val permissionService: PermissionService,
     ) {
-        fun getAllSnippets(): List<SnippetDto> {
-            return repository.findAll().map { snippet ->
-                val content = fetchSnippetContent(snippet.id!!)
-                translate(snippet, content)
+        fun getSnippet(
+            id: Long,
+            jwt: Jwt,
+        ): SnippetDto {
+            val canRead = permissionService.canRead(jwt, id).block() ?: false
+            if (!canRead) {
+                throw ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied")
             }
-        }
-
-        fun getSnippet(id: Long): SnippetDto {
             val snippet =
                 repository.findSnippetById(id)
                     ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Snippet not found")
             val content = fetchSnippetContent(id)
-            return translate(snippet, content)
+            return translate(snippet, content, "READ")
         }
 
-        fun createSnippet(snippetDto: CreateSnippetDto): SnippetDto {
+        fun createSnippet(
+            snippetDto: CreateSnippetDto,
+            jwt: Jwt,
+        ): SnippetDto {
             validateSnippetContent(snippetDto.content, snippetDto.version)
 
             val snippet = translate(snippetDto)
             val savedSnippet = repository.save(snippet)
             assetService.createAsset("snippets", savedSnippet.id.toString(), snippetDto.content).block()
-
-            return translate(savedSnippet, snippetDto.content)
+            permissionService.addPermission(jwt, savedSnippet.id!!, "OWNER").block()
+            return translate(savedSnippet, snippetDto.content, "OWNER")
         }
 
         fun createFromFile(
             snippetDto: CreateSnippetFileDto,
             file: MultipartFile,
+            jwt: Jwt,
         ): SnippetDto {
             val content = file.inputStream.bufferedReader(Charsets.UTF_8).use { it.readText() }
             validateSnippetContent(content, snippetDto.version)
-
             val snippet = translate(snippetDto)
             val savedSnippet = repository.save(snippet)
             assetService.createAsset("snippets", savedSnippet.id.toString(), content).block()
-
-            return translate(savedSnippet, content)
+            savedSnippet.id?.let { permissionService.addPermission(jwt, it, "owner") }
+            return translate(savedSnippet, content, "owner")
         }
 
         fun editSnippet(
             snippetDto: CreateSnippetDto,
             id: Long,
+            jwt: Jwt,
         ): SnippetDto {
             validateSnippetContent(snippetDto.content, snippetDto.version)
+            val canModify = permissionService.canModify(jwt, id).block() ?: false
+            if (!canModify) {
+                throw ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied")
+            }
 
             assetService.createAsset("snippets", id.toString(), snippetDto.content).block()
 
@@ -80,16 +91,21 @@ class SnippetService
                     ),
                 )
 
-            return translate(savedSnippet, snippetDto.content)
+            return translate(savedSnippet, snippetDto.content, "OWNER")
         }
 
         fun editFromFile(
             snippetDto: CreateSnippetFileDto,
             file: MultipartFile,
             id: Long,
+            jwt: Jwt,
         ): SnippetDto {
             val content = file.inputStream.bufferedReader(Charsets.UTF_8).use { it.readText() }
             validateSnippetContent(content, snippetDto.version)
+            val canModify = permissionService.canModify(jwt, id).block() ?: false
+            if (!canModify) {
+                throw ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied")
+            }
 
             val savedSnippet =
                 repository.save(
@@ -104,12 +120,21 @@ class SnippetService
                 )
             assetService.createAsset("snippets", savedSnippet.id.toString(), content).block()
 
-            return translate(savedSnippet, content)
+            return translate(savedSnippet, content, "OWNER")
         }
 
         @Transactional
-        fun deleteSnippet(id: Long) {
+        fun deleteSnippet(
+            id: Long,
+            jwt: Jwt,
+        ) {
+            val canModify = permissionService.canModify(jwt, id).block() ?: false
+            if (!canModify) {
+                throw ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied")
+            }
             repository.deleteById(id)
+            permissionService.removePermission(jwt, id, "OWNER").block()
+            assetService.deleteAsset("snippets", id.toString()).block()
         }
 
         private fun fetchSnippetContent(id: Long): String {
@@ -126,6 +151,24 @@ class SnippetService
                     ?: throw ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Error validating snippet")
             if (!validation.ok) {
                 throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid snippet content")
+            }
+        }
+
+        fun getSnippetsByUser(jwt: Jwt): List<SnippetDto> {
+            return permissionService.getAllSnippetPermissions(jwt).toIterable().map {
+                    permission ->
+                val snippet =
+                    repository.findSnippetById(permission.snippetId)
+                        ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Snippet not found")
+                val content = fetchSnippetContent(permission.snippetId)
+                translate(snippet, content, permission.permissionType)
+            }
+        }
+
+        fun getAllSnippets(): List<SnippetDto> {
+            return repository.findAll().map {
+                val content = fetchSnippetContent(it.id!!)
+                translate(it, content, "READ")
             }
         }
 
@@ -150,6 +193,7 @@ class SnippetService
         private fun translate(
             snippet: Snippet,
             content: String,
+            permission: String,
         ) = SnippetDto(
             id = snippet.id,
             name = snippet.name,
@@ -158,5 +202,6 @@ class SnippetService
             version = snippet.version,
             content = content,
             extension = snippet.extension,
+            permission = permission,
         )
     }
